Revert "[kyo-stats-machine] make the classpath contract true; generic FFI rel…"
This reverts commit 625349e45b9d6764d3ee587fbeceda48add8b0ad.
