+++ kyo-core/shared/src/test/scala/kyo/StatTest.scala
@@ -2,118 +2,6 @@ package kyo
 
 class StatTest extends kyo.test.Test[Any]:
 
-    // Cross-platform on purpose: the safe tier is what a caller reaches for on every target, and the whole
-    // point of this method is that reading a histogram back no longer costs an AllowUnsafe import. There is
-    // deliberately no `import AllowUnsafe.embrace.danger` in this leaf.
-    "Histogram.summary reads the distribution through the safe tier, with no AllowUnsafe at the call site" in {
-        val histogram = Stat.initScope("stattest-safe-summary").initHistogram("latency")
-        for
-            _     <- histogram.observe(10L)
-            _     <- histogram.observe(20L)
-            _     <- histogram.observe(30L)
-            first <- histogram.summary
-            again <- histogram.summary
-        yield
-            assert(first.count == 3)
-            assert(first.sum == 60.0)
-            assert(first.min <= 10.0)
-            assert(first.max >= 30.0)
-            // Non-destructive, unlike Counter.get: buckets and sum describe the whole lifetime, so a
-            // dashboard can poll this without disturbing an exporter reading the same instrument.
-            assert(again.count == first.count)
-            assert(again.sum == first.sum)
-        end for
-    }
-
-    // The hazard these exist for: the registering accessors are first-writer-wins, so a consumer that
-    // calls initCounter for a metric it does not own registers its own instrument at that path and
-    // permanently shadows the producer's value. Reading someone else's metric has to ask, not mint. No
-    // AllowUnsafe here either: this is the safe tier's answer, and StatsRegistry's own find* stay the
-    // unsafe-tier equivalents an exporter uses on the hot path.
-    "Stat.findCounter returns the producer's counter and reads it through the safe tier" in {
-        val scope    = Stat.initScope("stattest-find-counter")
-        val produced = scope.initCounter("hits")
-        for
-            _     <- produced.inc
-            _     <- produced.add(4L)
-            found <- scope.findCounter("hits")
-            seen <- found match
-                case Present(c) => c.get
-                case Absent     => Kyo.lift(-1L)
-        yield
-            assert(found.isDefined)
-            assert(seen == 5L)
-        end for
-    }
-
-    "Stat.findCounter never registers, so a misspelled name stays Absent and the real one is untouched" in {
-        val scope = Stat.initScope("stattest-find-typo")
-        val real  = scope.initCounter("requests")
-        for
-            _       <- real.inc
-            missing <- scope.findCounter("requsts")
-            // The discriminating part: if the miss had registered, this second look would find the thing
-            // the first look created, and the producer's counter would have been shadowed.
-            again <- scope.findCounter("requsts")
-            found <- scope.findCounter("requests")
-            value <- found match
-                case Present(c) => c.get
-                case Absent     => Kyo.lift(-1L)
-        yield
-            assert(missing.isEmpty)
-            assert(again.isEmpty)
-            assert(value == 1L)
-        end for
-    }
-
-    "Stat.findGauge reads the producer's thunk rather than replacing it, which initGauge would" in {
-        val scope = Stat.initScope("stattest-find-gauge")
-        val _     = scope.initGauge("temperature")(42.0)
-        for
-            found <- scope.findGauge("temperature")
-            v <- found match
-                case Present(g) => g.collect
-                case Absent     => Kyo.lift(-1.0)
-        yield
-            assert(found.isDefined)
-            assert(v == 42.0)
-        end for
-    }
-
-    "Stat.findHistogram reads the distribution non-destructively" in {
-        val scope    = Stat.initScope("stattest-find-histogram")
-        val produced = scope.initHistogram("latency")
-        for
-            _     <- produced.observe(10L)
-            _     <- produced.observe(20L)
-            found <- scope.findHistogram("latency")
-            first <- found match
-                case Present(h) => h.summary.map(s => (s.count, s.sum))
-                case Absent     => Kyo.lift((-1L, -1.0))
-            second <- found match
-                case Present(h) => h.summary.map(s => (s.count, s.sum))
-                case Absent     => Kyo.lift((-1L, -1.0))
-        yield
-            assert(first._1 == 2L)
-            assert(first._2 == 30.0)
-            // Re-reading sees the same lifetime totals: a draining read would report 0 here, which is what
-            // makes this safe for a dashboard polling alongside an exporter.
-            assert(second == first)
-        end for
-    }
-
-    "Histogram.summary sees observations made through the same registered path" in {
-        // The singleton-per-path dedup is what lets a reader and a writer share an instrument; the safe read
-        // has to observe that, not a fresh empty one.
-        val writer = Stat.initScope("stattest-safe-summary-shared").initHistogram("latency")
-        val reader = Stat.initScope("stattest-safe-summary-shared").initHistogram("latency")
-        for
-            _ <- writer.observe(5.0)
-            s <- reader.summary
-        yield assert(s.count == 1)
-        end for
-    }
-
     "scope".onlyJvm in {
         import AllowUnsafe.embrace.danger
         val stat         = Stat.initScope("test1")
