+++ kyo-ffi/plugin/src/main/scala/kyo/ffi/sbt/FfiLibrary.scala
@@ -51,17 +51,6 @@ import sbt._
   *   additional linker flags (appended to global `ffiLinkFlags`).
   * @param staticLink
   *   when true, statically link third-party libs for this library.
-  * @param osTargets
-  *   the OS names (`"linux"`, `"darwin"`, `"windows"`) whose shared library this
-  *   library is built and bundled for; empty (the default) means every OS. A
-  *   binding that only exists to be called on one OS names it here, so a build on
-  *   any other host neither compiles a shared library for it nor claims it in the
-  *   native manifest. Without it, C that is `#ifdef`-guarded to same-signature
-  *   stubs off its OS still compiles everywhere, and the build ships a working-
-  *   looking artifact for a platform that can never call it while the platform
-  *   that can may have none at all. `"linux-musl"` resolves the `"linux"` key.
-  *   Scala Native is unaffected: it compiles every declared C source into the
-  *   binary on every OS, which is what keeps the stub symbols resolvable there.
   * @param dependsOn
   *   ids of other libraries this library depends on. Used to topologically
   *   order C compilation so a library that `#include`s another's header (or
@@ -80,18 +69,8 @@ final case class FfiLibrary(
     linkFlags: Seq[String] = Nil,
     staticLink: Boolean = false,
     dependsOn: Seq[String] = Nil,
-    compilerByOs: Map[String, String] = Map.empty,
-    osTargets: Seq[String] = Nil
+    compilerByOs: Map[String, String] = Map.empty
 ) {
-
-    /** Whether this library's shared library is built and bundled on `os` (the resolved TARGET os).
-      * True for every OS when `osTargets` is empty, which is the default and every existing library's
-      * behaviour. `linux-musl` resolves the `linux` key.
-      */
-    def buildsOn(os: String): Boolean = {
-        val key = if (os == "linux-musl") "linux" else os
-        osTargets.isEmpty || osTargets.contains(key)
-    }
 
     /** Effective link libraries for the OS being built: the always-on `linkLibs`
       * plus the entry in `linkLibsByOs` for `os` (the resolved TARGET os,
