+++ kyo-ffi/plugin/src/main/scala/kyo/ffi/sbt/KyoFfiPlugin.scala
@@ -68,32 +68,6 @@ object KyoFfiPlugin extends AutoPlugin {
                 "<os>-<arch> directory, parsed from that suffix. Every id must be declared in ffiLibraries / " +
                 "ffiLibraryId. None (the default) merges nothing."
         )
-        val ffiReleasePlatforms = settingKey[Seq[String]](
-            "The platform keys a RELEASE of this project must ship a native for, derived from each " +
-                "library's own osTargets rather than named by hand: a darwin-only library requires exactly " +
-                "the darwin keys, one with no osTargets requires every supported key. `ffiPackagingCheckAll` " +
-                "uses this so the release workflow names no module, and a new FFI module is covered the day " +
-                "it is added. Read-only in practice; ffiRequiredPlatforms is what the check consumes and " +
-                "stays empty for developer builds."
-        )
-        val ffiPrebuiltPool = settingKey[Option[File]](
-            "A directory of prebuilt natives SHARED by several projects, from which this one takes only the " +
-                "files whose library id it declares and ignores the rest. Same naming convention as " +
-                "ffiPrebuiltDir. This is the release topology: producers on each OS compile every FFI module " +
-                "and upload one artifact per os-arch, the publish host downloads them all into one directory, " +
-                "and each project picks its own out of it, so adding an FFI module needs no release-workflow " +
-                "change. ffiPrebuiltDir stays strict (an undeclared id there is a mistake, because that " +
-                "directory was pointed at this project deliberately); the pool is shared by construction, so " +
-                "another module's native in it is expected rather than an error. None (the default) merges " +
-                "nothing."
-        )
-        val ffiRequiredPlatforms = settingKey[Seq[String]](
-            "Platform keys (`<os>-<arch>`) that `ffiPackagingCheck` requires a native for, per bundled " +
-                "library. Empty (the default) means a build packages whatever this host produced, which is " +
-                "right for a developer build; a release sets the full supported set so a producer that did " +
-                "not run fails the publish instead of shipping a jar with a hole in it."
-        )
-
         val ffiStubLibraries = settingKey[Seq[String]](
             "Library ids whose declared C sources are a placeholder rather than the real binding (e.g. a TLS " +
                 "shim compiled with no vendored library staged). Recorded as 'stub' in the library-state " +
@@ -111,16 +85,11 @@ object KyoFfiPlugin extends AutoPlugin {
         val FfiLibrary = kyo.ffi.sbt.FfiLibrary
 
         // Tasks
-        val ffiGenerate   = taskKey[Seq[File]]("Generate platform-specific impl sources from bindings.")
-        val ffiCompile    = taskKey[Seq[File]]("Compile C sources into a platform-native shared library.")
-        val ffiPackage    = taskKey[Seq[File]]("Copy the compiled library into META-INF/native/ in resources.")
-        val ffiClean      = taskKey[Unit]("Clean generated sources + compiled libs.")
-        val ffiCiWorkflow = taskKey[File]("Emit a starter .github/workflows/ffi-native.yml template.")
-        val ffiPackagingCheck = taskKey[Unit](
-            "Validate the staged native resource tree: every artifact's binary format and extension match " +
-                "the platform directory it sits in, and every ffiRequiredPlatforms entry has a native for " +
-                "every bundled library."
-        )
+        val ffiGenerate          = taskKey[Seq[File]]("Generate platform-specific impl sources from bindings.")
+        val ffiCompile           = taskKey[Seq[File]]("Compile C sources into a platform-native shared library.")
+        val ffiPackage           = taskKey[Seq[File]]("Copy the compiled library into META-INF/native/ in resources.")
+        val ffiClean             = taskKey[Unit]("Clean generated sources + compiled libs.")
+        val ffiCiWorkflow        = taskKey[File]("Emit a starter .github/workflows/ffi-native.yml template.")
         val ffiNpmBundleTemplate = taskKey[File]("Emit package.json pinning koffi to the supported range (Scala.js consumers).")
         val ffiNativeLinkingOptions = taskKey[Seq[String]](
             "Scala Native linkingOptions for ffiLibraries: the static-folded SYSTEM link libs " +
@@ -211,103 +180,6 @@ object KyoFfiPlugin extends AutoPlugin {
 
     import autoImport._
 
-    /** Compiles the natives of EVERY project that enables this plugin, so a release producer runs one
-      * command instead of naming each FFI module.
-      *
-      * The workflow's producer jobs previously carried a hand-written block per module: one `ffiCompile`
-      * invocation, one collect step and one artifact per module, duplicated for the next one. Adding an FFI
-      * module then meant editing the release workflow in four places, which is how a module ends up
-      * published with no native for a platform it needs. The plugin already knows which projects have FFI
-      * libraries, so it is the thing that should enumerate them.
-      *
-      * A project whose `ffiLibraries` resolve to nothing compilable on this host is skipped rather than
-      * failed: a darwin-only shim on a Linux producer is the expected case, not an error.
-      */
-    private def ffiCompileAllCommand: Command = Command.command("ffiCompileAll") { state =>
-        val extracted = Project.extract(state)
-        // A project enables this plugin exactly when its settings carry ffiLibraries, which the plugin
-        // defaults for every project it is enabled on and nothing else defines.
-        val projects = extracted.structure.allProjectRefs.filter { ref =>
-            (ref / ffiLibraries).get(extracted.structure.data).isDefined
-        }
-        if (projects.isEmpty) {
-            state.log.warn("[kyo-ffi-plugin] ffiCompileAll: no project enables KyoFfiPlugin; nothing to compile.")
-            state
-        } else {
-            state.log.info(
-                s"[kyo-ffi-plugin] ffiCompileAll: ${projects.size} project(s): " +
-                    projects.map(_.project).sorted.mkString(", ")
-            )
-            val (finalState, failed) =
-                projects.foldLeft((state, Seq.empty[String])) { case ((st, bad), ref) =>
-                    Project.runTask(ref / ffiCompile, st) match {
-                        case Some((next, Value(_))) => (next, bad)
-                        case Some((next, Inc(_)))   => (next, bad :+ ref.project)
-                        case None                   => (st, bad :+ ref.project)
-                    }
-                }
-            if (failed.nonEmpty)
-                sys.error(
-                    s"[kyo-ffi-plugin] ffiCompileAll: ffiCompile failed for " +
-                        s"${failed.sorted.mkString(", ")}. See the errors above."
-                )
-            finalState
-        }
-    }
-
-    /** Runs `ffiPackagingCheck` on every FFI project, with the required platforms DERIVED per library
-      * from its own `osTargets` rather than named in the release workflow.
-      *
-      * `ffiRequiredPlatforms` has to stay empty for a developer build, which packages whatever this host
-      * produced, and be the full set for a release, which must fail rather than ship a jar with a hole in
-      * it. Naming that set per module in the workflow is what made the workflow grow a block per module,
-      * and it is knowledge the build already has: a library declaring `osTargets = Seq("darwin")` requires
-      * exactly the darwin platform keys, and one declaring nothing requires all supported keys.
-      *
-      * So the release runs one command, every module is covered, and a new FFI module is checked the day
-      * it is added without anyone remembering to add it here.
-      */
-    private def ffiPackagingCheckAllCommand: Command = Command.command("ffiPackagingCheckAll") { state =>
-        val extracted = Project.extract(state)
-        val projects = extracted.structure.allProjectRefs.filter { ref =>
-            (ref / ffiLibraries).get(extracted.structure.data).isDefined
-        }
-        val (finalState, failed) =
-            projects.foldLeft((state, Seq.empty[String])) { case ((st, bad), ref) =>
-                val ex           = Project.extract(st)
-                val requiredKeys = (ref / ffiReleasePlatforms).get(ex.structure.data).getOrElse(Nil)
-                if (requiredKeys.isEmpty) (st, bad)
-                else {
-                    st.log.info(
-                        s"[kyo-ffi-plugin] ffiPackagingCheckAll: ${ref.project} requires " +
-                            requiredKeys.mkString("[", ", ", "]")
-                    )
-                    val withReq = ex.appendWithoutSession(
-                        Seq(ref / ffiRequiredPlatforms := requiredKeys),
-                        st
-                    )
-                    Project.runTask(ref / ffiPackagingCheck, withReq) match {
-                        case Some((next, Value(_))) => (next, bad)
-                        case Some((next, Inc(_)))   => (next, bad :+ ref.project)
-                        case None                   => (st, bad :+ ref.project)
-                    }
-                }
-            }
-        // Every failing project is reported, not just the first, so one release run names everything that
-        // has to be fixed rather than one thing per attempt.
-        if (failed.nonEmpty)
-            sys.error(
-                s"[kyo-ffi-plugin] ffiPackagingCheckAll: the packaging check failed for " +
-                    s"${failed.sorted.mkString(", ")}. See the errors above."
-            )
-        finalState
-    }
-
-    override lazy val globalSettings: Seq[Setting[?]] = Seq(
-        commands += ffiCompileAllCommand,
-        commands += ffiPackagingCheckAllCommand
-    )
-
     override lazy val projectSettings: Seq[Setting[?]] = Seq(
         ffiLibraryId := "kyo_ffi",
         ffiCSources := {
@@ -335,19 +207,9 @@ object KyoFfiPlugin extends AutoPlugin {
         // Unset means the host: every producer resolves the target through
         // CCompiler.resolveTargetOsArch, so a build that never sets this behaves exactly as before.
         // The system property lets a CI matrix leg name its target without a build edit.
-        ffiTargetOsArch := sys.props.get("kyo.ffi.targetOsArch"),
-        ffiPrebuiltDir  := None,
-        ffiPrebuiltPool := None,
-        ffiReleasePlatforms := {
-            val libs = ffiLibrariesResolved.value
-            // buildsOn is the same predicate ffiCompile gates on, so what a release requires cannot drift
-            // from what a producer would actually build.
-            libs.filter(_.cSources.nonEmpty).flatMap { lib =>
-                CCompiler.supportedOsArchTags.filter(key => lib.buildsOn(CCompiler.parseOsArch(key)._1))
-            }.distinct.sorted
-        },
-        ffiRequiredPlatforms := Nil,
-        ffiStubLibraries     := Nil,
+        ffiTargetOsArch  := sys.props.get("kyo.ffi.targetOsArch"),
+        ffiPrebuiltDir   := None,
+        ffiStubLibraries := Nil,
         // Common system libraries that bindings may reference without the plugin
         // producing or packaging an artifact for them. Users can extend this list
         // to whitelist additional system-provided libraries.
@@ -688,17 +550,7 @@ object KyoFfiPlugin extends AutoPlugin {
                     // omitted from the macOS / Windows command, and IS included when a build targets
                     // Linux from elsewhere.
                     libs.zipWithIndex.flatMap { case (lib, idx) =>
-                        if (lib.cSources.nonEmpty && !lib.buildsOn(targetOs)) {
-                            // Declared for other OSes only. Its C compiles here (the sources are
-                            // `#ifdef`-guarded to same-signature stubs off their OS) but the artifact could
-                            // only ever be loaded on an OS this is not, so building it would ship a
-                            // working-looking native for a platform that can never call it.
-                            log.info(
-                                s"[kyo-ffi-plugin] ffiCompile: ${lib.id} targets ${lib.osTargets.mkString(", ")}; " +
-                                    s"skipping on $targetOs."
-                            )
-                            Nil
-                        } else if (lib.cSources.isEmpty) {
+                        if (lib.cSources.isEmpty) {
                             // Declared with no C sources: an INTENTIONALLY-ABSENT native, not a failed
                             // build. The library-state manifest records it as `absent` so a packaging
                             // completeness check does not read the missing artifact as a broken leg.
@@ -757,8 +609,7 @@ object KyoFfiPlugin extends AutoPlugin {
         },
 
         // ffiPackage: explicit task that copies artifacts; same body as the resource generator below.
-        ffiPackage        := ffiPackagedNatives.value,
-        ffiPackagingCheck := ffiPackagingCheckTask.value,
+        ffiPackage := ffiPackagedNatives.value,
 
         // Copy compiled artifacts (and any staged prebuilts) into the resource tree automatically.
         // On Native this is a no-op (ffiCompile returns Nil). On JVM and JS the
@@ -1032,22 +883,6 @@ object KyoFfiPlugin extends AutoPlugin {
       * `ffiPrebuiltDir`, including its own; the dedicated producer staged BoringSSL for that os-arch
       * while the publish host may not, so the downloaded prebuilt is the authoritative artifact.
       */
-    /** Every prebuilt this project stages: the strict `ffiPrebuiltDir` plus its OWN entries from the shared
-      * `ffiPrebuiltPool`.
-      *
-      * Every consumer of staged prebuilts has to use this, not `prebuiltNatives(ffiPrebuiltDir.value)`. Under
-      * the pool topology a module's natives arrive in the shared directory, so a generator reading only the
-      * strict one reports them missing: the library-state manifest would call a staged library `absent`, and
-      * the native manifest would omit the platform it is filed under, which is the exact shape of the defect
-      * that shipped a macOS shim no Mac could load.
-      */
-    private def stagedPrebuilts(
-        prebuiltDir: Option[File],
-        poolDir: Option[File],
-        declared: Set[String]
-    ): Seq[PrebuiltNative] =
-        prebuiltNatives(prebuiltDir) ++ prebuiltNatives(poolDir).filter(p => declared.contains(p.libraryId))
-
     private def ffiPackagedNatives: Def.Initialize[Task[Seq[File]]] = Def.task {
         val log         = streams.value.log
         val compiled    = ffiCompile.value
@@ -1055,27 +890,12 @@ object KyoFfiPlugin extends AutoPlugin {
         val platform    = ffiTargetPlatform.value
         val libs        = ffiLibrariesResolved.value
         val prebuiltDir = ffiPrebuiltDir.value
-        val poolDir     = ffiPrebuiltPool.value
+        val prebuilt    = prebuiltNatives(prebuiltDir)
 
         val (targetOs, targetArch) = CCompiler.resolveTargetOsArch(ffiTargetOsArch.value)
 
-        val declared = libs.map(_.id).toSet
-
-        // The pool is shared with other FFI modules, so take only this project's own ids and leave the
-        // rest alone; the strict directory below still rejects anything undeclared. A file that parses
-        // but belongs to another module is the normal case here, not a mistake.
-        val fromPool = prebuiltNatives(poolDir).filter(p => declared.contains(p.libraryId))
-        poolDir.foreach { d =>
-            val mine = fromPool.map(_.file.getName).sorted
-            log.info(
-                s"[kyo-ffi-plugin] ffiPrebuiltPool ${d.getAbsolutePath}: taking " +
-                    (if (mine.isEmpty) "nothing (no native here declares one of this project's ids)"
-                     else mine.mkString(", "))
-            )
-        }
-
-        val prebuilt   = stagedPrebuilts(prebuiltDir, poolDir, declared)
-        val undeclared = prebuiltNatives(prebuiltDir).filterNot(p => declared.contains(p.libraryId))
+        val declared   = libs.map(_.id).toSet
+        val undeclared = prebuilt.filterNot(p => declared.contains(p.libraryId))
         if (undeclared.nonEmpty)
             sys.error(
                 s"[kyo-ffi-plugin] ffiPrebuiltDir stages natives for undeclared library ids: " +
@@ -1109,70 +929,6 @@ object KyoFfiPlugin extends AutoPlugin {
                     s"${prebuilt.map(p => s"${p.os}-${p.arch}").distinct.sorted.mkString(", ")}."
             )
         local ++ staged
-    }
-
-    /** Validate what this build actually staged, before it becomes a published artifact.
-      *
-      * Two failures, both of which have shipped:
-      *
-      *   - an artifact under a platform key it cannot be loaded on. The layout, the file name and the
-      *     extension can all agree while the bytes inside are for another OS, so the check reads the
-      *     object-file magic, which is the one thing that cannot be renamed.
-      *   - a platform with no artifact at all. A release builds each OS on its own runner and folds the
-      *     results together; if one producer does not run, nothing in a per-host build notices, and the jar
-      *     ships with a hole exactly where the library was needed. `ffiRequiredPlatforms` is the release's
-      *     declaration of the set that must be complete, and it is empty by default so a developer build
-      *     still packages whatever its own host produced.
-      *
-      * No-op on Native, which links its C into the binary and stages no shared library at all.
-      */
-    private def ffiPackagingCheckTask: Def.Initialize[Task[Unit]] = Def.task {
-        val log      = streams.value.log
-        val staged   = ffiPackagedNatives.value
-        val platform = ffiTargetPlatform.value
-        val required = ffiRequiredPlatforms.value
-        val module   = name.value
-        if (platform == "Native") ()
-        else {
-            // Every staged file sits directly under its own `<os>-<arch>` directory, for both the JVM
-            // (META-INF/native/...) and the JS (kyo-ffi/native/...) layouts.
-            val byPlatform = staged.map(f => (f.getParentFile.getName, f))
-
-            val mismatches = byPlatform.flatMap { case (key, file) => NativeArtifactFormat.mismatch(key, file) }
-            if (mismatches.nonEmpty)
-                sys.error(
-                    s"[kyo-ffi-plugin] $module: ${mismatches.size} staged native(s) do not match the platform " +
-                        s"directory they are in: ${mismatches.sorted.mkString("; ")}."
-                )
-
-            if (required.nonEmpty) {
-                val ids = byPlatform.flatMap { case (_, f) => NativeArtifactFormat.libraryIdOf(f.getName) }.distinct.sorted
-                val have = byPlatform.flatMap { case (key, f) =>
-                    NativeArtifactFormat.libraryIdOf(f.getName).map(id => (id, key))
-                }.toSet
-                val missing =
-                    for {
-                        id  <- ids
-                        key <- required
-                        if !have.contains((id, key))
-                    } yield s"$id has no native for $key"
-                if (missing.nonEmpty)
-                    sys.error(
-                        s"[kyo-ffi-plugin] $module: the staged native set is incomplete for the required " +
-                            s"platforms ${required.sorted.mkString("[", ", ", "]")}: ${missing.sorted.mkString("; ")}. " +
-                            s"Stage the missing producer's artifacts through ffiPrebuiltDir, or drop the platform " +
-                            s"from ffiRequiredPlatforms."
-                    )
-                log.info(
-                    s"[kyo-ffi-plugin] $module: ffiPackagingCheck passed, ${ids.size} librar(y/ies) present for " +
-                        s"${required.sorted.mkString(", ")}."
-                )
-            } else
-                log.info(
-                    s"[kyo-ffi-plugin] $module: ffiPackagingCheck passed, ${staged.size} staged native(s) match " +
-                        s"their platform directories (no ffiRequiredPlatforms set)."
-                )
-        }
     }
 
     /** Native-only resource generator: copy each library's C sources into
@@ -1364,21 +1120,18 @@ object KyoFfiPlugin extends AutoPlugin {
       * `native` id with no file in the layout is precisely the broken-leg case a check should reject.
       */
     private def ffiLibraryStateManifestGenerator: Def.Initialize[Task[Seq[File]]] = Def.task {
-        val platform = ffiTargetPlatform.value
-        val libs     = ffiLibrariesResolved.value
-        val stubs    = ffiStubLibraries.value.toSet
-        val prebuilt =
-            stagedPrebuilts(ffiPrebuiltDir.value, ffiPrebuiltPool.value, libs.map(_.id).toSet).map(_.libraryId).toSet
+        val platform   = ffiTargetPlatform.value
+        val libs       = ffiLibrariesResolved.value
+        val stubs      = ffiStubLibraries.value.toSet
+        val prebuilt   = prebuiltNatives(ffiPrebuiltDir.value).map(_.libraryId).toSet
         val resManaged = (Compile / resourceManaged).value
         val moduleName = name.value
-        val targetOs   = CCompiler.resolveTargetOsArch(ffiTargetOsArch.value)._1
         if (platform == "Native") Seq.empty[File]
         else {
             val states = libs.map { lib =>
                 val state =
-                    if (lib.cSources.nonEmpty && lib.buildsOn(targetOs)) {
-                        if (stubs.contains(lib.id)) "stub" else "native"
-                    } else if (prebuilt.contains(lib.id)) "prebuilt"
+                    if (lib.cSources.nonEmpty) { if (stubs.contains(lib.id)) "stub" else "native" }
+                    else if (prebuilt.contains(lib.id)) "prebuilt"
                     else "absent"
                 s"${lib.id}=$state"
             }
@@ -1433,7 +1186,7 @@ object KyoFfiPlugin extends AutoPlugin {
         val _                      = ffiGenerate.value
         val platform               = ffiTargetPlatform.value
         val libs                   = ffiLibrariesResolved.value
-        val prebuilt               = stagedPrebuilts(ffiPrebuiltDir.value, ffiPrebuiltPool.value, libs.map(_.id).toSet)
+        val prebuilt               = prebuiltNatives(ffiPrebuiltDir.value)
         val resManaged             = (Compile / resourceManaged).value
         val moduleName             = name.value
         val projTarget             = target.value
@@ -1443,8 +1196,7 @@ object KyoFfiPlugin extends AutoPlugin {
         else {
             val targetTag = s"$targetOs-$targetArch"
             val idBlocks = libs.sortBy(_.id).flatMap { lib =>
-                val local =
-                    if (lib.cSources.nonEmpty && lib.buildsOn(targetOs)) Set(targetTag) else Set.empty[String]
+                val local     = if (lib.cSources.nonEmpty) Set(targetTag) else Set.empty[String]
                 val staged    = prebuilt.filter(_.libraryId == lib.id).map(p => s"${p.os}-${p.arch}").toSet
                 val platforms = (local ++ staged).toSeq.sorted
                 Seq(
