+++ kyo-stats-machine/shared/src/main/scala/kyo/stats/machine/LinuxDisk.scala
@@ -218,7 +218,7 @@ private[machine] object LinuxDisk:
                 store.cell.total.set(blocks * frsize)
                 store.cell.free.observe(bavail * frsize)
             end if
-        catch case ex: Throwable if Machine.degradable(ex) => ()
+        catch case ex: Throwable if scala.util.control.NonFatal(ex) => ()
     end statvfsInto
 
 end LinuxDisk
