+++ kyo-stats-machine/shared/src/main/scala/kyo/stats/machine/Machine.scala
@@ -57,53 +57,4 @@ private[kyo] object Machine:
         def close()(using AllowUnsafe): Unit     = ()
     end NullMachine
 
-    /** Whether a failure degrades the family that raised it, instead of propagating out of the reader.
-      *
-      * `scala.util.control.NonFatal` alone is the wrong predicate here, and it is wrong for exactly the
-      * failure this module exists to absorb. A native library that is missing or unloadable surfaces as
-      * `ExceptionInInitializerError` the first time the generated binding's class initializes, as
-      * `NoClassDefFoundError` on every later touch of that now-poisoned class, and as `UnsatisfiedLinkError`
-      * when the library resolves but a symbol does not. All three are `LinkageError`s, which `NonFatal`
-      * excludes, so a guard written with `NonFatal` lets the one error it was written for escape.
-      *
-      * A genuinely fatal error (`OutOfMemoryError`, `StackOverflowError`, an interrupt) still propagates:
-      * degrading on those would hide a process-level problem behind a missing metric.
-      */
-    private[machine] def degradable(ex: Throwable): Boolean =
-        scala.util.control.NonFatal(ex) || ex.isInstanceOf[LinkageError]
-
-    /** Already-reported degradations, keyed by the label passed to `reportDegraded`. */
-    private val reported = new java.util.concurrent.ConcurrentHashMap[String, java.lang.Boolean]
-
-    /** Reports that `what` degraded to absent, at most once per label for the process lifetime.
-      *
-      * A telemetry layer that cannot read the host must say so once, then stay quiet: the failure repeats on
-      * every tick, so reporting each occurrence would write a line per second into an application that never
-      * asked for host metrics. One `warn` line carries the cause's own message, which for a load failure is
-      * the searched paths, the missing symbol and the override property, and the same line is emitted on
-      * every platform, including JS where the failure was previously invisible.
-      *
-      * The cause is rendered into the message rather than handed to the logger as a `Throwable`: the module's
-      * contract is a quietly missing metric, never a stack trace in an unrelated application's stderr.
-      *
-      * @return
-      *   true when this call was the one that reported, false when the label had already been reported.
-      */
-    private[machine] def reportDegraded(what: String, ex: Throwable)(using AllowUnsafe): Boolean =
-        if reported.putIfAbsent(what, java.lang.Boolean.TRUE) ne null then false
-        else
-            given Frame = Frame.internal
-            Log.live.unsafe.warn(degradedMessage(what, ex))
-            true
-
-    /** The one line a degraded family reports.
-      *
-      * It carries the cause's own rendering, which for a native-load failure is the searched paths, the
-      * missing symbol and the override property the loader documents. That is what makes the same line
-      * useful on every platform: the JVM's `FfiLoadError` message is a model of what a good load failure
-      * says, and this puts it in front of a JS user, where the failure used to be entirely silent.
-      */
-    private[machine] def degradedMessage(what: String, ex: Throwable): String =
-        s"kyo-stats-machine: $what is unavailable on this host, so its metrics stay absent. Cause: $ex"
-
 end Machine
