+++ kyo-stats-machine/shared/src/test/scala/demo/MachineStatsDemo.scala
@@ -1,17 +1,16 @@
 package demo
 
 import kyo.*
-import kyo.stats.machine.MachineMetrics
 import kyo.stats.machine.MachineRegistrySnapshot
 
 /** Host observability with zero setup: put kyo-stats-machine on the classpath and read machine.* metrics.
   *
   * An operator running a kyo service wants CPU, memory, disk, and load telemetry without wiring any
-  * monitoring API into their code. This demo is that operator's program: it never calls a Machine method,
-  * never starts a sampler by hand, and never names a metric. Being a `kyo.KyoApp` is the whole of it: the
-  * entrypoint reaches `kyo.Stat` before the app's own code runs, which is enough for the classpath-present
-  * kyo-stats-machine module to auto-load its exporter factory, start the host sampler, and begin feeding the
-  * shared `kyo.Stat` registry. After a few sampler ticks the demo reads the registry through `MachineRegistrySnapshot`, which
+  * monitoring API into their code. This demo is that operator's program: it never calls a Machine method
+  * and never starts a sampler by hand. It only touches `kyo.Stat` (the class every metrics-emitting kyo app
+  * already touches), which is enough for the classpath-present kyo-stats-machine module to auto-load its
+  * exporter factory, start the once-per-second host sampler, and begin feeding the shared `kyo.Stat`
+  * registry. After a few sampler ticks the demo reads the registry through `MachineRegistrySnapshot`, which
   * enumerates it exactly as `OTLPMetricsExporter` does, and reports the machine.* families it found with real
   * values sampled off THIS host.
   *
@@ -29,7 +28,7 @@ import kyo.stats.machine.MachineRegistrySnapshot
   * own run is how you watch the opt-out suppress the sampler.
   *
   * Demonstrates:
-  *   - classpath-presence auto-load: running as a `KyoApp` starts the host sampler, with no user API call
+  *   - classpath-presence auto-load: touching `kyo.Stat` alone starts the host sampler, with no user API call
   *   - the machine.* kyo.Stat metric taxonomy (cpu / memory / swap / disk / load) populated from the real host
   *   - reading the kyo.Stat registry through the exporter-facing store snapshot (MachineRegistrySnapshot)
   *   - graceful degradation: cgroup / PSI families are Absent on a non-Linux host, never faked
@@ -53,17 +52,18 @@ object MachineStatsDemo:
 
     private val ticksToObserve = 3
 
-    /** The operator flow: let the sampler run a few ticks, then snapshot the registry the way an exporter
-      * reads it.
-      *
-      * There is no user action at all. `kyo.KyoApp`'s entrypoint reaches `kyo.Stat` before running this code,
-      * which runs the service-loader scan, which constructs the classpath-present `MachineStatFactory`, which
-      * starts the sampler. No Machine method is called and no metric is named; this is what an application
-      * that merely has the module on its classpath already does. (A host that is not a `KyoApp` calls
-      * `Stat.activate()` once instead; see the module README.)
+    /** The operator flow. Touch `kyo.Stat` to trigger auto-load (no Machine API call), let the sampler run a
+      * few ticks, then snapshot the registry the way an exporter reads it.
       */
     def flow(using Frame): Report < (Async & Abort[Throwable]) =
         for
+            // The ONLY user action: reference kyo.Stat so object Stat's class-init eager service-loader scan
+            // runs, which constructs the classpath-present MachineStatFactory, which starts the sampler.
+            // No Machine method is called; this is what a metrics-only app already does.
+            _ <- Sync.defer {
+                import AllowUnsafe.embrace.danger
+                Stat.initScope("app").initCounter("touch", "forces Stat class-init").unsafe.add(0)
+            }
             // Wait past the first sampler tick (which only records the cumulative baseline) plus a couple more,
             // so histograms have observations and CPU counters carry a real cumulative advance.
             _ <- Async.sleep((ticksToObserve + 1).seconds)
@@ -97,55 +97,13 @@ object MachineStatsDemo:
         )
     end report
 
-    /** The metric families a healthy host must report, named through the module's own key constants rather
-      * than string literals.
-      *
-      * This is the set an audit of the released artifact recorded as present on a working host, and it is what
-      * makes the check sensitive to a PARTIAL loss. A host whose native reader is broken outright reports one
-      * family (`cpu.cores`, which comes from the JDK's processor count and needs no native), so any check at
-      * all catches that; a regression that drops only swap, or only two of the four cpu rates, is invisible
-      * unless the whole set is named.
-      *
-      * Two families are required everywhere except Windows, because that host does not report them and the
-      * readers say so rather than faking a value. It has no load-average concept, so the three `load.*`
-      * gauges are absent; and `GlobalMemoryStatusEx` exposes no free-versus-available distinction, so
-      * `memory.free` is never written rather than repeating the available figure under a second label.
-      * Requiring either there would demand a number the module deliberately does not produce. Disk is checked
-      * separately, by mount, since its family names are host-specific.
-      */
-    def requiredKeys(os: String): Chunk[MachineMetrics.Key] =
-        val everywhere = Chunk(
-            MachineMetrics.cpuTotalRate,
-            MachineMetrics.cpuUserRate,
-            MachineMetrics.cpuSystemRate,
-            MachineMetrics.cpuIdleRate,
-            MachineMetrics.cpuCores,
-            MachineMetrics.memoryTotal,
-            MachineMetrics.memoryAvailable,
-            MachineMetrics.swapTotal,
-            MachineMetrics.swapFree
-        )
-        if os == "Windows" then everywhere
-        else
-            everywhere.concat(Chunk(
-                MachineMetrics.memoryFree,
-                MachineMetrics.loadOne,
-                MachineMetrics.loadFive,
-                MachineMetrics.loadFifteen
-            ))
-        end if
-    end requiredKeys
-
     /** Acceptance check for a Report. Returns Absent when the report proves auto-load fed real host metrics into
       * kyo.Stat, Present(reason) otherwise. Every threshold traces to a physical fact about a running host,
       * not to whatever the sampler happened to produce.
       */
     def validate(r: Report): Maybe[String] =
-        val missing = requiredKeys(r.os).filterNot(k => r.sampled.exists(_.path == k.dotted))
         if r.sampled.isEmpty then
             Present("no machine.* metrics in the registry: auto-load did not start the sampler")
-        else if missing.nonEmpty then
-            Present(s"machine.* families missing on ${r.os}: ${missing.map(_.dotted).mkString(", ")}")
         else if r.memoryTotalBytes.isEmpty then
             Present("machine.memory.total absent: the sampler did not observe host memory")
         // A real host reports more than 256 MiB of total RAM; the memory.total gauge holds that byte count.
@@ -168,7 +126,6 @@ object MachineStatsDemo:
         else if r.os != "Linux" && r.pressurePresent then
             Present(s"machine.pressure.* present on a non-Linux host (${r.os}): graceful degradation violated (PSI must be Absent)")
         else Absent
-        end if
     end validate
 
     /** Raised by the runnable entry point when `validate` rejects the report, so the process exits non-zero
@@ -182,9 +139,9 @@ end MachineStatsDemo
   * validation passes and non-zero (via [[MachineStatsDemo.ValidationFailed]]) when it does not, so a failed
   * run on your own classpath surfaces through the exit code rather than only in the printed output.
   *
-  * Auto-load is triggered by the `KyoApp` entrypoint, with nothing in the flow itself asking for it; the
-  * opt-out is read once at that moment, so running under `KYO_MACHINE_DISABLED=true` yields an empty snapshot
-  * and a validation failure (the demo's proof that the opt-out suppresses the sampler).
+  * Auto-load is triggered by the flow's single `kyo.Stat` touch; the opt-out is read once at that touch, so
+  * running under `KYO_MACHINE_DISABLED=true` yields an empty snapshot and a validation failure (the demo's
+  * proof that the opt-out suppresses the sampler).
   */
 object MachineStatsDemoApp extends KyoApp:
     run {
