+++ kyo-system/js-wasm/src/test/scala/kyo/internal/SystemPlatformSpecificJsWasmTest.scala
@@ -2,7 +2,6 @@ package kyo.internal
 
 import kyo.*
 import kyo.AllowUnsafe.embrace.danger
-import scala.scalajs.js.annotation.JSImport
 import scala.scalajs.js as sjs
 
 class SystemPlatformSpecificJsWasmTest extends kyo.test.Test[Any]:
@@ -46,33 +45,4 @@ class SystemPlatformSpecificJsWasmTest extends kyo.test.Test[Any]:
         }
     }
 
-    "availableProcessors" - {
-
-        "reports the host's logical processor count, not the Scala.js stub" in {
-            // The control: Scala.js's javalib stub answers 1 on every host, which is what
-            // machine.cpu.cores reported on a 12-core Mac. It is also the denominator every per-core
-            // normalisation divides by, and the input the CPU histogram's buckets are derived from.
-            assert(Runtime.getRuntime.availableProcessors() == 1)
-            assert(SystemPlatformSpecific.availableProcessors() == NodeOs.availableParallelism())
-        }
-
-        "is stable across calls" in {
-            assert(SystemPlatformSpecific.availableProcessors() == SystemPlatformSpecific.availableProcessors())
-        }
-
-        "falls back to a positive count with no process global" in {
-            assert(withoutProcessGlobal(SystemPlatformSpecific.availableProcessors()) >= 1)
-        }
-    }
-
 end SystemPlatformSpecificJsWasmTest
-
-/** Node's `os` module, read through the `node:` specifier so the facade links under both the CommonJS (js)
-  * and ESModule (wasm) backends. `availableParallelism` is the count Node itself reports as available to this
-  * process, the independent oracle this suite checks `availableProcessors` against.
-  */
-@sjs.native
-@JSImport("node:os", JSImport.Namespace)
-private object NodeOs extends sjs.Object:
-    def availableParallelism(): Int = sjs.native
-end NodeOs
